Introduce PrefixTree class
Our approach for automated argument completion will require the ability
to determine which argument strings subsume a given word (the one to
complete), or, put differently, the ability to find the options which
the given word is a prefix of.
A popular data structure that helps in making such a decision is a trie
or prefix tree.
This change introduces a a PrefixTree class which will later be used to
find all completion candidates for the given argument.