Add functionality for iterating over words
A prefix tree (in our implementation) is useful for finding a node whose
textual representation is a super set of the prefix we searched for.
However, it does not directly answer the question of which words the
given word can be completed to.
To retrieve a list of words this patch introduces the iterWords()
function. This function iterates over all known registered words that
can be found from a given node.