Properly escape '--'
With our approach of passing in the $COMP_WORDS array as a list of words
instead of as a single string we run into problems if one of the words
is the string '--'. This string is treated special by the ArgumentParser
class in that it marks the remaining arguments as positional. This
behavior in turn results in the --_complete option not receiving all
words.
To fix this issue this change escapes any '--' string after the
--_complete option in a way that the ArgumentParser does not care about.
In the handler for the --_complete action we revert the escaping and can
subsequently attempt to complete the '--' string.