Add sub parser support
In its current state we are able to complete arguments in a single
argument parser. However, by using the add_subparsers() method it is
possible to create sub parsers which represent sub commands and should
in turn provide completion functionality.
This change adds support for arbitrary deep nesting of argument parsers.
We achieve the correct behavior by representing each sub parser as
another prefix tree which is embedded in the parent prefix tree.