Add functionality to determine number of arguments to expect
Argument completion should not only work at the level of program options
but also for arguments to those options (if any). Such arguments are
typically created by providing a non-zero nargs value to the
ArgumentParser's add_argument method or, implicitly, by specifying a
certain type of action.
In order to deduce the number of arguments to expect, this change adds
functionality to decode the 'action' as well as the 'nargs' parameter.