Add support for positional arguments
The logic implemented thus far solely deals with keyword arguments, that
is, arguments beginning with one or two dashes. Equally important,
however, are positional arguments -- arbitrary strings (not beginning
with a dash) whose semantic depends on their order relative to each
other rather than a keyword. There are two types of positional
arguments: those specified at the level of an individual parser and
those belonging to a given keyword argument.
This patch adjusts the parser to support both types of positional
arguments. It uses the previously introduced functionality to decode the
nargs parameter properly to infer how many positionals to expect.