Replace PrefixTree class with a namedtuple
The prefix tree initially devised to ease the finding of prefixes turned
out to be less suitable for the increasing number of use-cases we try to
support. Positional arguments, for example, cannot be represented easily
in such a tree. Even more, the program starts more and more to evolve
around this class and how to interface with its objects than to focus on
the logic that is to be implemented.
To alleviate this problem, this change replaces the PrefixTree class
with a simple tuple representing positional and keyword arguments.