Use a list for storing the positionals
Previously we dealt with parser-level positional arguments by storing a
single <min, max> tuple representing all of them. This approach is
insufficient once we want to complete individual positional arguments,
as they might be of different types and, hence, have different
completions.
Although support for such completions is not yet done, in order to be in
a better position to support such a change in the future, this patch
converts the positional arguments from a single aggregate value to a
list of <min, max> pairs each representing one positional argument type
as specified by the user.