Correctly handle multiple positional arguments
Our completion breaks whenever a positional argument actually comprises
multiple arguments by means of the nargs option. An example of this
practice is the acceptance of an arbitrary number of files as the 'cat'
program does.
This change adjusts the complete function to be able to deal with
arbitrary nargs values.