Suppress _complete agument in namespace
The CompletingArgumentParser class is meant as a drop-in replacement for
argparse's ArgumentParser class with the additional feature of providing
completion support. As such, the interface the former provides should
not differ from the original interface in any way (barring additions).
However, the namespace as produced by a CompletingArgumentParser
contains a special member, _complete, because of the added --_complete
option. Depending on how the parser's produced namespace is used, this
additional member has the potential to break compatibility.
This change removes this field from the namespace by adding the
--_complete argument with a suppressed default value.