Add argument group support
The ArgumentParser class support grouping of arguments. This grouping
has no real semantic effect aside from customizing the help text
produced. Unfortunately, as group affiliation is established by adding
the respective arguments to a group object as opposed to a parser
object, our parser with completion support is not aware of them and,
thus, no completions will be provided.
This change addresses the problem by intercepting the creation of such a
group object and overwriting its add_argument method to properly
register a completion if desired.