Add completion support for parse_known_args
Since we only customized the argument parser's parse_args method, the
parse_known_args method does not work properly in conjunction with
the --_complete option as a double dash is not escaped properly and,
thus, will screw up parsing.
This patch addresses this problem by overwriting the parse_known_args
method as well.