Introduce Argument class
We currently treat the minimum and maximum number of positional and
keyword arguments as a simple tuple. Because of that, we always need to
work with the unpacked (individual) values, which was found to
complicate the code unnecessarily.
This change introduces a new class, Argument, that is a named tuple
grouping the minimum and maximum values together and making them
accessible as member. By doing so we increase readability, especially of
the complete() function.