Add choice completion support
Not only do we want to complete options themselves but also the
arguments for those options. Only in very few cases it is actually
possible to infer the potential completions for an individual argument
because a string argument, for example, has no information attached as
to what it represents. However, if an argument supports 'choices',
completing this argument becomes trivial: everything that could be
completed to one of those choices is a valid candidate.
This change adds support for completing arguments that are choices.