Correctly reset positional argument state for each subparser
When we encounter a sub parser we should not only reset the current
state of expected keyword arguments but also that of positional
arguments. We are effectively dealing with a new completely new parser
not having anything in common with the previous one.
This change fixes a problem where we kept the previous positional
argument state, potentially resulting in a sub parser not handling
positional arguments correctly.