Add support for custom completers
If possible completions are not known at "compile" time, the 'choices'
argument is not suitable for advising possible completions. For such
cases, we need a way to infer the completions for a string at run time.
Two approaches to accomplish this task have been identified. Either we
can somehow make use of the 'type' and/or 'action' arguments to
add_argument() and infer completion from them. For instance, we could
check if a specified type sports a "complete" method and, if so, use it
to infer possible completions. Another approach is to pass in a
dedicated argument solely used for generating completions. Both
approaches have pros and cons. Encoding completion information in
existing parameters preserves full compatibility with argparse's
ArgumentParser (if done properly), at an increased complexity of the
result (in terms of the implementation because parameters are overloaded
but also potentially at the client side because additional logic is
required to fit in completions). An extra argument breaks compatibility
when going back from the parser with completion support to the standard
ArgumentParser because it might not understand this additional argument
and error out.
This patch follows the second approach and introduces a new argument to
the add_argument function. This argument is a completion function that
accepts a word to complete and produces the possible completions.