Supply 'parser' and 'values' to each completer
Completers cannot always perform a useful completion using solely global
state. Sometimes it is important to know the value of other arguments
and perform completion based on them.
To that end, this change passes the parser used for the completion as
well as the full set of arguments ('values') to the complete function
which in turn hands them through to the completer invocation. By doing
so we give completer the chance to parse the values themselves in order
to determine the value of certain arguments.