Factor out CompletingArgumentParser.complete method
The CompleteAction's __call__ method does at least two things: it
unescapes and "parses" the passed in values and then it performs a
completion based on the index and the words of interest that it
extracted beforehand.
This change refactors the method and moves the completion part into a
new method within the CompletingArgumentParser class. This approach
makes sense because a lot of the parser's members are used in the
completion. Note that this change does not introduce any change in
functionality.