Do not rely on ArgumentParser implementing parse_args with parse_known_args
The CompletingArgumentParser supports not only the parse_args method but
also its counterpart parse_known_args. However, the implementation
required the former to be implemented by invoking the latter within
argparse's ArgumentParser class. This fact is an implementation detail
and must not be relied upon but at the time no better way was found that
avoids multiple double-dash escapes taking place.
With this change we fix up this imperfection by putting proper machinery
in place to skip multiple escapes.