Introduce completePath() function for file completion
Depending on the context in which applications are developed, in many
cases an argument may actually refer to a file, i.e., represent its
path. Completing files is not as straight forward as one would expect
and reimplementing such common functionality everytime is not
beneficial.
To that end, this change introduces a new completer, completePath, that
can be used to complete file paths relative to the current working
directory.