Register completePath() for arguments of type argparse.FileType
The argparse's ArgumentParser supports a special type, FileType, for
arguments that causes the parser to automatically open the supplied file
making it ready for use after a parse_args() call.
Knowing that only files are valid inputs for arguments of this type,
completion becomes feasible. This change hooks up the completePath()
completer for arguments of this type in order to support completion out
of the box.