Provide reStructuredText formatted README file
Platforms such as PyPI interpret the long description attribute of the
setup.py in order to display a project page. Unfortunately, the
expectation is that this file is formatted as reStructuredText and not
using markdown.
This change adds a README.rst file that is properly formatted and
instructs the setup.py file to read it when asked for the long
description. Our markdown formatted README.md will stay the reference
and the rst file is simply generated from it.